+++ CHANGELOG.md
@@ -1,2 +1,4 @@
+- Allow `entryFileGlob` to be an array or a string
+
 # v1.0.1 - **2017-01-30**
 # v1.0.0 - **2017-01-30**